Added 'Edit goal' form
Let’s be honest, this isn’t the best way of doing this at all. And I
have one issue with the validation. But sure, this is it!

In the future, this won’t be a form, but will be directly editable by
hovering over the goal title and such.

The issue with validation is issue #16 - the edit goal button remains
enabled even if the input is invalid.